requirements.txt
beautifulsoup4==4.10.0
faiss-cpu==1.5.3
fasttext-wheel==0.9.2
Flask==1.1.4
Flask_Table==0.5.0
gensim==4.0.1
langdetect==1.0.8
nltk==3.6.1
numpy==1.20.2
pandarallel==1.5.2
pandas==1.2.3
python_bcrypt==0.3.2
pywikibot==6.3.0
requests==2.25.1
scikit_learn==1.0.1
scipy==1.6.2
torch==1.8.1
mwparserfromhell>=0.5.0
markupsafe==2.0.1
pytest==7.0.1
configparser==5.2.0
